backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==4.0.1

# Embeddings and ML
sentence-transformers==2.3.1
torch==2.1.2
numpy==1.26.3

# LLM integrations
openai==1.10.0
anthropic==0.18.1
httpx==0.26.0

# Text processing
langchain-community==0.0.17
langchain-text-splitters==0.0.1
tiktoken==0.5.2

# CORS
python-cors==1.0.0
